requirements-dev.txt
asgiref==3.2.3
Django==3.0.2
django-environ==0.4.5
djangorestframework==3.11.0
requests==2.24.0
pycpfcnpj==1.5.1


boto3==1.16.16
botocore==1.19.16
celery==5.0.2
redis==3.5.3
watchtower==1.0.0
